major refactor. pitch split into an abstract class BasePitch. With Pitch (horizontal) and VerticalPitch (vertical) inheriting from the BasePitch class. This depreciates the orientation parameter. The layout parameter is depreciated in favour of the matplotlib style nrows and ncols. View is depreciated and replaced with a half boolean (True for half sized pitch). The stripey background for grass pitch and plotting methods are in progress and not included in this commit. The reason for this refactor is to remove much of the if/else code.
